requirements.txt
numpy
PyYAML
pandas
Pillow
scipy==1.7.1
scikit-learn==0.24.2